Refactor guard and prisoner to redux toolkit
